watch_dog/requirements_watchdog.txt
watchdog==5.0.3
requests==2.32.3
python-dotenv==1.0.1
pika==1.3.2

# lib pro extrakci pdf -> txt
pdfminer.six
mysql-connector-python

# konvert pdf na image
pdf2image
pytesseract